Apply suggestions from code review

+++ hooks/terraform_providers_lock.sh
@@ -16,7 +16,7 @@ function main {
   # JFYI: suppress color for `terraform providers lock` is N/A`
 
   # Cleanup in case lockfile was not cleaned in previous run of common::terraform_init function
-  rmdir "$PARALLELISM_FALLBACK_LOCK_DIR"
+  [ -d "$PARALLELISM_FALLBACK_LOCK_DIR" ] && rmdir "$PARALLELISM_FALLBACK_LOCK_DIR"
   # shellcheck disable=SC2153 # False positive
   common::per_dir_hook "$HOOK_ID" "${#ARGS[@]}" "${ARGS[@]}" "${FILES[@]}"
 }
